Add setting to allow "hidden" tabs.
BuddyPress 2.4.0 allows user to visit tabs that are not included in the
hub navigation. This change takes advantage of that and allows a tab to
exist but be excluded from a hub’s navigation. Meant to be used very
sparingly, it might be useful in very specific instances, for example
when persistent navigation to those pages exists somewhere else like in
the footer text.